Clase Pedido terminada

+++ app/src/main/java/com/example/adminportatil/cebanctortillas/Producto.java
@@ -1,12 +1,13 @@
 package com.example.adminportatil.cebanctortillas;
 
 
+
 public class Producto {
-    private int precio;
+    private double precio;
     private int cantidad;
     private String descripcion;
 
-    Producto (int precio, int cantidad, String descripcion){
+    Producto (double precio, int cantidad, String descripcion){
         this.precio= precio;
         this.cantidad=cantidad;
         this.descripcion= descripcion;
